requirements.txt
discord.py
asyncio
aiohttp
websockets
chardet
datetime
PyNaCl
emoji
regex
Pillow
vk_api
bs4
requests